venator/requirements.txt
asgiref==3.3.1
blinker==1.4
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
dataclasses==0.8
Flask==1.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kaitaistruct==0.9
ldap3==2.8.1
MarkupSafe==1.1.1
mitmproxy==5.3.0
msgpack==1.0.2
passlib==1.7.4
protobuf==3.13.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
PySocks==1.7.1
python-dotenv==0.16.0
redis==3.5.3
requests==2.25.1
rq==1.7.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
selenium==3.141.0
selenium-wire==4.2.2
six==1.15.0
sortedcontainers==2.2.2
tornado==6.1
urllib3==1.26.4
urwid==2.1.2
Werkzeug==1.0.1
wsproto==0.15.0
zstandard==0.14.1
